Refactored "admin cell" CLI.
- All commands refactored into profile specific plugins.
