+++ entry_points.txt
@@ -22,9 +22,15 @@ krb5keytab = treadmill_aws.cli.krb5keytab
 
 [treadmill.cli.admin]
 aws = treadmill_aws.cli.admin.aws
-cell = treadmill_aws.cli.admin.cell
 krb5keytab = treadmill_aws.cli.admin.krb5keytab
 ipa = treadmill_aws.cli.admin.ipa
+
+[treadmill.cli.admin.cell.aws]
+configure-apps = treadmill_aws.cli.admin.cell.configure_apps
+configure-monitors = treadmill_aws.cli.admin.cell.configure_monitors
+configure-appgroups = treadmill_aws.cli.admin.cell.configure_appgroups
+configure-dns = treadmill_aws.cli.admin.cell.configure_dns
+restart-apps = treadmill_aws.cli.admin.cell.restart_apps
 
 [treadmill_aws.cli.aws]
 image = treadmill_aws.cli.aws.image
