+++ lib/python/treadmill_aws/cli/admin/cell/templates/identity-groups
